+++ packages/components-web/src/components/LandingCuisineCard/index.ts
@@ -1,0 +1,2 @@
+export { LandingCuisineCard, default } from './LandingCuisineCard';
+export type { LandingCuisineCardProps } from './LandingCuisineCard';